backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
pymysql>=1.0.3
pydantic>=1.10.7
pydantic[email]>=1.10.7
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
requests>=2.28.2
pyjwt>=2.6.0
python-dotenv
